pyproject.toml
[tool.poetry]
name = "sqlalchemy-databricks"
version = "0.1.0"
description = "SQLAlchemy Dialect for Databricks"
authors = ["flynn <crf204@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/crflynn/sqlalchemy-databricks"
repository = "https://github.com/crflynn/sqlalchemy-databricks"
documentation = "https://github.com/crflynn/sqlalchemy-databricks"
keywords = ["databricks", "sqlalchemy", "dialect"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Database",
    "Topic :: Database :: Front-Ends",
]
include = [
    { path="CHANGELOG.md", format="sdist" },
    { path="LICENSE.txt", format="sdist" },
]

[tool.poetry.plugins."sqlalchemy.dialects"]
"databricks.connector" = "sqlalchemy_databricks:DatabricksDialect"

[tool.poetry.dependencies]
python = "^3.7"
databricks-sql-connector = "^0"
SQLAlchemy = "^1"
PyHive = "^0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.2"
pytest = "^6.2.5"
black = "^21.12b0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.github
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
)
'''

[tool.isort]
force_single_line = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120
